+++ packages/Hyperlink/src/Hyperlink.test.tsx
@@ -2,61 +2,56 @@
  * @jest-environment jsdom
  */
 import React from 'react';
-import { render } from '@testing-library/react';
-import { shallow } from 'enzyme';
+import { render, screen } from '@testing-library/react';
 import Plus from '@igloo-ui/icons/dist/Plus';
 
 import Hyperlink from './Hyperlink';
 
 const setUp = (props = {}) => {
-  const component = shallow(
+  return render(
     <Hyperlink dataTest="ids-link" {...props}>
       Click me
     </Hyperlink>
   );
-  return component;
 };
 
 describe('Hyperlink Component', () => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let component: any;
-  beforeEach(() => {
-    component = setUp();
-  });
-
   test('It should render without errors', () => {
-    const wrapper = component.find('.ids-link');
-    expect(wrapper.length).toBe(1);
+    setUp();
+    const wrapper = screen.getByTestId('ids-link');
+    expect(wrapper).toBeInTheDocument();
   });
 
   test('It should render a hyperlink with icon', (): void => {
-    render(<Hyperlink iconLeading={<Plus size="small" />}>Click me</Hyperlink>);
-    const svg = document.getElementsByTagName('svg');
-    expect(svg.length).toBe(1);
+    const { container } = setUp({
+      iconLeading: <Plus size="small" />,
+    });
+    const svg = container.querySelector('svg');
+    expect(svg).toBeInTheDocument();
   });
 
   test('It should render a primary appearance', (): void => {
-    const hyperlink = setUp();
-    expect(hyperlink).toMatchSnapshot();
+    const { asFragment } = setUp();
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a secondary appearance', (): void => {
-    const hyperlink = setUp({ appearance: 'secondary' });
-    expect(hyperlink).toMatchSnapshot();
+    const { asFragment } = setUp({ appearance: 'secondary' });
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a premium appearance', (): void => {
-    const hyperlink = setUp({ appearance: 'premium' });
-    expect(hyperlink).toMatchSnapshot();
+    const { asFragment } = setUp({ appearance: 'premium' });
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a danger appearance', (): void => {
-    const hyperlink = setUp({ appearance: 'danger' });
-    expect(hyperlink).toMatchSnapshot();
+    const { asFragment } = setUp({ appearance: 'danger' });
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a ghost appearance', (): void => {
-    const hyperlink = setUp({ appearance: 'ghost' });
-    expect(hyperlink).toMatchSnapshot();
+    const { asFragment } = setUp({ appearance: 'ghost' });
+    expect(asFragment()).toMatchSnapshot();
   });
 });