+++ packages/Input/src/Input.test.tsx
@@ -2,66 +2,80 @@
  * @jest-environment jsdom
  */
 import React from 'react';
-import { shallow } from 'enzyme';
+import { render, screen } from '@testing-library/react';
 
 import Input from './Input';
 import Search from '@igloo-ui/icons/dist/Search';
 
 const setUp = (props = {}) => {
-  const component = shallow(<Input dataTest="ids-input" {...props} />);
-  return component;
+  return render(<Input dataTest="ids-input" {...props} />);
 };
 
 describe('Input', () => {
   test('It should render without errors', (): void => {
-    const input = setUp();
-    const wrapper = input.find('.ids-input');
-    expect(wrapper.length).toBe(1);
+    setUp();
+    const wrapper = screen.getByTestId('ids-input');
+    expect(wrapper).toBeInTheDocument();
   });
 
   test('It should render a small size', (): void => {
-    const input = setUp({ isCompact: true });
-    const wrapper = input.find('.ids-input--compact');
-    expect(wrapper.length).toBe(1);
+    setUp({ isCompact: true });
+    const wrapper = screen.getByTestId('ids-input');
+    expect(wrapper).toHaveClass('ids-input--compact');
   });
 
   test('It should render a readonly', (): void => {
-    const input = setUp({ disabled: true });
-    const wrapper = input.find('.ids-input--disabled');
-    expect(wrapper.length).toBe(1);
+    setUp({ disabled: true });
+    const wrapper = screen.getByTestId('ids-input');
+    expect(wrapper).toHaveClass('ids-input--disabled');
   });
 
   test('It should render a snapshot', (): void => {
-    const input = setUp();
-    expect(input).toMatchSnapshot();
+    const { asFragment } = setUp();
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render an email type', (): void => {
-    const input = setUp({ value: 'hashbrown@potatocatinc.com' });
-    expect(input).toMatchSnapshot();
+    const { asFragment } = setUp({
+      type: 'email',
+      value: 'hashbrown@potatocatinc.com',
+    });
+    expect(screen.getByTestId('ids-input').getAttribute('type')).toEqual(
+      'email'
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a text type', (): void => {
-    const input = setUp({ type: 'test', value: 'potat' });
-    expect(input).toMatchSnapshot();
+    const { asFragment } = setUp({ type: 'text', value: 'potato' });
+    expect(screen.getByTestId('ids-input').getAttribute('type')).toEqual(
+      'text'
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a password type', (): void => {
-    const input = setUp({
+    const { asFragment } = setUp({
       type: 'password',
       value: 'this is a great password!',
     });
-    expect(input).toMatchSnapshot();
+    expect(screen.getByTestId('ids-input').getAttribute('type')).toEqual(
+      'password'
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a number type', (): void => {
-    const input = setUp({ type: 'number', value: 12345 });
-    expect(input).toMatchSnapshot();
+    const { asFragment } = setUp({ type: 'number', value: 12345 });
+    expect(screen.getByTestId('ids-input').getAttribute('type')).toEqual(
+      'number'
+    );
+    expect(asFragment()).toMatchSnapshot();
   });
 
   test('It should render a input with prefix icon', (): void => {
-    const input = setUp({ prefixIcon: <Search size="small" /> });
-    const wrapper = input.find('.ids-input--prefixIcon');
-    expect(wrapper.length).toBe(1);
+    setUp({ prefixIcon: <Search size="small" /> });
+    const wrapper = screen.getByTestId('ids-input');
+    expect(wrapper).toHaveClass('ids-input--prefixIcon');
   });
 });