+++ packages/IconButton/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @igloo-ui/icon-button
 
+## 0.2.11
+
+### Patch Changes
+
+- d971143: All tests were updated to use testing library and not enzyme
+- Updated dependencies [d971143]
+  - @igloo-ui/button@0.3.5
+
 ## 0.2.10
 
 ### Patch Changes